+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- Added support for VS Code task API. Now users can add their custom task definitions to the `tasks.json` file. See README for details.
+
+### Changed
+- Reintroduced source code display in the flamegraph.
+
 ## [0.3.5] - 2025-02-24
 ### Fixed
 - Fixed issue where profiles were not rendered in Jupyter cells with CRLF line endings